Per review comments
