+++ include/openthread/types.h
@@ -1088,6 +1088,37 @@ typedef void (OTCALL *otDeviceAvailabilityChangedCallback)(bool aAdded, const GU
 
 #endif // OTDLL
 
+/**
+ * Log levels.
+ *
+ * Implimentation note: Log Levels are done as defines so that they
+ * may be used on the embedded system side as compile time defines
+ * and thus tested with #if/#else/#endif constructs.
+ */
+typedef uint8_t otLogLevel;
+
+/**
+ * This enum represents log regions.
+ *
+ */
+typedef enum otLogRegion
+{
+    OT_LOG_REGION_API      = 1,  ///< OpenThread API
+    OT_LOG_REGION_MLE      = 2,  ///< MLE
+    OT_LOG_REGION_ARP      = 3,  ///< EID-to-RLOC mapping.
+    OT_LOG_REGION_NET_DATA = 4,  ///< Network Data
+    OT_LOG_REGION_ICMP     = 5,  ///< ICMPv6
+    OT_LOG_REGION_IP6      = 6,  ///< IPv6
+    OT_LOG_REGION_MAC      = 7,  ///< IEEE 802.15.4 MAC
+    OT_LOG_REGION_MEM      = 8,  ///< Memory
+    OT_LOG_REGION_NCP      = 9,  ///< NCP
+    OT_LOG_REGION_MESH_COP = 10, ///< Mesh Commissioning Protocol
+    OT_LOG_REGION_NET_DIAG = 11, ///< Network Diagnostic
+    OT_LOG_REGION_PLATFORM = 12, ///< Platform
+    OT_LOG_REGION_COAP     = 13, ///< CoAP
+} otLogRegion;
+
+
 #ifdef __cplusplus
 }  // extern "C"
 #endif
